ChageColor Menu change opsition of footer.jsx
