+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
-import { Geist, Geist_Mono, Kanit } from "next/font/google";
+import { Kanit } from "next/font/google";
 import "./globals.css";
-import Footer from "@/components/Footer";
+import Footer from "./components/Footer";
 import Navbar from "./components/Navbar/Navbars";
 
 const kanit = Kanit({
@@ -21,9 +21,9 @@ export default function RootLayout({
 }>) {
     return (
         <html lang="en">
-            <body className={`${kanit.className} antialiased`}>
+            <body className={`${kanit.className} antialiased bg-black`}>
                 <Navbar />
-                {children}
+                <div className="min-h-screen">{children}</div>
                 <Footer />
             </body>
         </html>
