+++ package.json
@@ -9,11 +9,12 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@emotion/react": "^11.14.0",
     "@heroicons/react": "^2.2.0",
     "antd": "^5.24.0",
+    "dayjs": "^1.11.13",
     "framer-motion": "^12.4.2",
     "hamburger-react": "^2.5.2",
-    "dayjs": "^1.11.13",
     "next": "15.1.7",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
